chore: rename connection form to profile form
