+++ packages/sshnp_gui/lib/src/presentation/screens/new_connection_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:sshnp_gui/src/presentation/widgets/new_connection_form.dart';
+import 'package:sshnp_gui/src/presentation/widgets/profile/new_profile_form.dart';
 
 import '../../utils/sizes.dart';
 import '../widgets/navigation/app_navigation_rail.dart';
@@ -32,7 +32,7 @@ class _NewConnectionScreenState extends State<NewConnectionScreen> {
                     style: Theme.of(context).textTheme.titleMedium,
                   ),
                   gapH10,
-                  const Expanded(child: NewConnectionForm())
+                  const Expanded(child: NewProfileForm())
                 ]),
               ),
             ),
